refactor: OutView 기능 분리
